Divided extension
