+++ partytion-app/PartytionApp/Presentation/Component/CustomUIColor.swift
@@ -1,0 +1,25 @@
+//
+//  CustomUIColor.swift
+//  PartytionApp
+//
+//  Created by tatsuya_oyanagi on 2019/10/04.
+//  Copyright © 2019 naruhiyo. All rights reserved.
+//
+
+import UIKit
+
+extension UIColor {
+    convenience init(_ red: Int, _ green: Int, _ blue: Int, _ alpha: Int = 255) {
+        let rgba = [red, green, blue, alpha].map { i -> CGFloat in
+            switch i {
+            case let i where i < 0:
+                return 0
+            case let i where i > 255:
+                return 1
+            default:
+                return CGFloat(i) / 255
+            }
+        }
+        self.init(red: rgba[0], green: rgba[1], blue: rgba[2], alpha: rgba[3])
+    }
+}
